Fix for Issue 59.
There was limited success in fixing this defect.

Because the Google Maps library tries to load before the page renders, the user is left waiting for that to timeout -- which can be anywhere from 20s to 1-2min (its the browser and operating system that decide, and it is not configurable by the end user).

However, on the plus side: when the browser gives up trying to get the maps library, the page shows: "TAMT is checking your internet connection." If there is no connection, the page shows: "It appears your internet connection is not working. Please make sure your TAMT VirtualBox is connected to the Internet, then reload the page".

When the user connects to the internet and reloads the page, the TAMT application appears.